Revision 508:
- Floppytreiber objektorientierter gemacht, basiert auf floppy_t
- flpydsk_refreshVolumeNames statt flpydsk_get_volumeName
- diskType_t mit Funktionspointern für Sector-Read/-Write für Abstraktion
- Vereinfachungen im Code

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@508 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/include/usb2_msd.h
@@ -29,7 +29,7 @@ uint8_t usbTransferBulkOnlyGetMaxLUN(uint32_t device, uint8_t numInterface);
 void usbSendSCSIcmd(uint32_t device, uint32_t interface, uint32_t endpointOut, uint32_t endpointIn, uint8_t SCSIcommand, uint32_t LBA, uint16_t TransferLength, usbBulkTransfer_t* bulkTransfer);
 
 void testMSD(uint8_t devAddr, partition_t* part);
-uint8_t usbRead(uint32_t sector, uint8_t* buffer);
+int32_t usbRead(uint32_t sector, uint8_t* buffer);
 
 void usbResetRecoveryMSD(uint32_t device, uint32_t Interface, uint32_t endpointOUT, uint32_t endpointIN);
